Moved common code to MAC file due to the fact freeBSD and openBSD is not currently supported in processes inventory
