+++ src/wazuh_modules/syscollector/sysInfoOpenBSD.cpp
@@ -87,3 +87,8 @@ std::string SysInfo::getSerialNumber() const
     spBuff.get()[len] = 0;
     return std::string{reinterpret_cast<const char*>(spBuff.get())};
 }
+
+nlohmann::json SysInfo::getProcessesInfo() const
+{
+    return {};
+}
