+++ src/wazuh_modules/syscollector/testtool/CMakeLists.txt
@@ -66,6 +66,7 @@ else()
 	    dl
 	    pthread
 		proc
+		gcov
 	)
 
 	if(SOLARIS)
